Started styling

+++ baymanagement/src/EditBay.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
 import './App.css';
 import axios from 'axios'
+import swal from 'sweetalert'
 
 class EditBay extends Component {
 	constructor() {
@@ -70,6 +71,7 @@ class EditBay extends Component {
 				console.log(res.data);
 				let newState = Object.assign({},this.state.masterBayInfo);
 				newState.message = res.data.message
+				newState.message == "edit successful" ? swal(`${newState.message}`, "", "success") : swal(`${newState.message}`, "", "error")
 				this.setState({
 					masterBayInfo: newState
 				});
@@ -77,6 +79,10 @@ class EditBay extends Component {
 				console.log(this.state.masterBayInfo);
 			})
 		console.log("finished hitting post");
+
+		
+		
+
 	}
 
 	deleteHandler(event) {
@@ -84,8 +90,27 @@ class EditBay extends Component {
 		let bId = this.props.match.params.id;
 		bId = parseInt(bId)
 		console.log(bId, "delete");
-		axios.delete(`http://localhost:8080/deleteMasterBay?id=${bId}`)
-		this.props.history.push('/?msg=deleted')
+
+		  swal({
+			title: "Are you sure?",
+			text: `MasterBay ${bId} will be deleted!`,
+			icon: "warning",
+			buttons: true,
+			dangerMode: true,
+		  })
+		  .then((willDelete) => {
+			if (willDelete) {
+			  swal(`MasterBay ${bId} has been deleted.`, {
+				icon: "success",
+			  });
+			  	axios.delete(`http://localhost:8080/deleteMasterBay?id=${bId}`)
+				this.props.history.push('/')
+			} else {
+			  swal("Delete cancelled");
+			}
+		  });
+
+
 	}
 
 	addSubHandler(event) {
@@ -100,6 +125,22 @@ class EditBay extends Component {
 		return (
 			// <BrowserRouter>
 				<div>
+
+					
+					<form className="bar" onSubmit={this.submitHandler}>
+						<label>
+							ID: MB
+							<input type="number" name="id" id="id"
+								value={this.state.masterBayInfo.id} disabled />
+						</label>
+						<br />
+
+						<label>
+							Height:
+						<input type="number" name="height" 
+						value={this.state.masterBayInfo.height} 
+						onChange={this.handleChangeHeight()} />
+
 					<h2>Editing MB{this.state.masterBayInfo.id}</h2>
 					<form className="bar" onSubmit={this.submitHandler}>
 						<label>
@@ -107,6 +148,7 @@ class EditBay extends Component {
 							<input type="text" name="length" 
 								value={this.state.masterBayInfo.length} 
 								onChange={this.handleChangeLength()} />
+
 						</label>
 						<br />
 
@@ -119,16 +161,23 @@ class EditBay extends Component {
 						<br />
 
 						<label>
+
+							Length:
+							<input type="number" name="length" 
+								value={this.state.masterBayInfo.length} 
+								onChange={this.handleChangeLength()} />
+
 							Height:
 						<input type="text" name="height" 
 						value={this.state.masterBayInfo.height} 
 						onChange={this.handleChangeHeight()} />
+
 						</label>
 						<br />
 
-						<h3>{this.state.masterBayInfo.message}</h3>
-						<div className="row">
-							<button className="btn btn-primary" 
+					
+						<div className="row ">
+							<button className="btn btn-primary custom-btn" 
 								type="submit">
 								Submit
 							</button>
@@ -136,20 +185,20 @@ class EditBay extends Component {
 					</form>
 
 					<div>
-						<button className="btn btn-primary" 
+						<button className="btn btn-primary custom-btn" 
 							type="submit" 
 							onClick={this.deleteHandler}>
 							Delete
 						</button>
 					</div>
 
-					<div className="row">
-						<button className="btn btn-primary" 
+					{/* <div className="row">
+						<button className="btn btn-primary custom-btn" 
 							type="submit" 
 							onClick={this.ManageSubs}>
 							Manage Sub Bays
 						</button>
-					</div>
+					</div> */}
 				</div>
 			// </BrowserRouter>
 		)
